Pytesting

+++ app/services/users_service.py
@@ -36,6 +36,8 @@ def put_user(new_data: schemas.UserCreate, db: Session, user_id: int):
 
 def delete_user(user_id: int, db: Session):
     user = db.query(models.User).filter(models.User.id == user_id).first()
+    if not user:
+        raise HTTPException(status_code=404, detail="user not found")
     db.delete(user)
     db.commit()
     return db.query(models.User).all()
